refactor(libsy): internalize fall-through routing
